Reviews GitHub action to use the cross-platform build.sh script.
Dotnet 5 (as well as core 3.1-) supports cross-compiling by
specifiaction of the RuntimeIdentifier argument. This allows to a single
instance build and package the whole multiplatform set.

It also takes advantage of the script that calls dotnet publish to place
the files in the right output directory for the NPM package.